working on combining correlation alignments to one script
