+++ pyspecProcScripts/correlation_alignment.py
@@ -11,7 +11,9 @@ def to_percent(y, position):
         return s + r'$\%$'
     else:
         return s + '%'
-def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_title='correlation alignment',ph1_selection=1,ph2_selection=1, sigma = 20,fl=None):
+def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',
+        fig_title='correlation alignment',ph1_selection=1,ph2_selection=0, 
+        shift_bounds=False, max_shift = 100., sigma=20.,direct='t2',fl=None):
     """
     Align transients collected with chunked phase cycling dimensions along an indirect
     dimension based on maximizing the correlation across all the transients and repeat
@@ -33,8 +35,15 @@ def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_titl
                     sets the tolerance limit for the alignment procedure
     ph1_selection:  int
                     index position of the coherence pathway in phase program 1
-    ph2_selection:  int
-                    index position of the coherence pathway in phase program 2
+    shift_bounds:   boolean
+                    keeps f_shift to be within a specified
+                    limit (upper and lower bounds given by max_shift)
+                    which should be around the location of the expected
+                    signal
+    max_shift:      float
+                    specifies the upper and lower bounds to the range over
+                    which f_shift will be taken from the correlation function.
+                    shift_bounds must be True
     sigma:          int
                     sigma value for the gaussian fitting. Related to the linewidth
                     of the given data.
@@ -52,8 +61,23 @@ def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_titl
                 the width of the Gaussian function used to frequency filter
                 the data in the calculation of the correlation function.
     """
-    ph1_len = len(s.getaxis('ph1'))
-    ph2_len = len(s.getaxis('ph2'))
+    if 'nScans' in s.dimlabels:
+        phcycdims = [j for j in s.dimlabels if j.startswith('ph')]
+        indirect = set(s.dimlabels)-set(phcycdims)-set([direct])
+        indirect = [j for j in s.dimlabels if j in indirect]
+        nScans_len = len(s.getaxis('nScans'))
+        s.smoosh(indirect)
+    if 'ph2' in s.dimlabels:
+        ODNP=False
+    else:
+        ODNP=True
+    if ODNP:
+        s.ift(['ph1'])
+        ph1_len = len(s.getaxis('ph1'))
+    else:
+        s.ift(['ph1','ph2'])
+        ph1_len = len(s.getaxis('ph1'))
+        ph2_len = len(s.getaxis('ph2'))
     N = ndshape(s)[indirect_dim]
     sig_energy = (abs(s)**2).data.sum().item() / N
     if fl is not None:
@@ -66,8 +90,12 @@ def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_titl
     energy_vals.append(this_E / sig_energy)
     last_E = None
     for_nu_center =s.C
-    for_nu_center.ft(['ph1','ph2'])
-    for_nu_center = for_nu_center['ph1',ph1_selection]['ph2',ph2_selection]
+    if ODNP:
+        for_nu_center.ft(['ph1'])
+        for_nu_center = for_nu_center['ph1',ph1_selection]
+    else:
+        for_nu_center.ft(['ph1','ph2'])
+        for_nu_center = for_nu_center['ph1',ph1_selection]['ph2',ph2_selection]
     nu_center = for_nu_center.mean(indirect_dim).C.argmax('t2')
     logging.info(strm("Center frequency", nu_center))
     for my_iter in range(100):
@@ -87,18 +115,27 @@ def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_titl
         s_copy *= exp(-(s_copy.fromaxis('t2')-nu_center)**2/(2*sigma**2))
         s_copy.ift('t2')
         s_copy2 = s.C
-        s_copy *= nddata(r_[1.,1.],'DeltaPh2')
-        s_copy *= nddata(r_[1.,1.,1.,1.],'DeltaPh1')
-        s_copy.setaxis('DeltaPh2','#')
-        s_copy.setaxis('DeltaPh1','#')
+        if ODNP:
+            s_copy *= nddata(r_[1.,1.,1.,1.],'DeltaPh1')
+            s_copy.setaxis('DeltaPh1','#')
+        else: 
+            s_copy *= nddata(r_[1.,1.],'DeltaPh2')
+            s_copy *= nddata(r_[1.,1.],'DeltaPh1')
+            s_copy.setaxis('DeltaPh2','#')
+            s_copy.setaxis('DeltaPh1','#')
         correl = s_copy * 0 
-        for ph1_index in range(ph1_len):
-            print("PH1 INDEX IS:",ph1_index)
-            s_copy['DeltaPh1',ph1_index] = s_copy['DeltaPh1',ph1_index].run(lambda x, 
-                axis=None: roll(x, ph1_index,axis=axis),'ph1')
-        for ph2_index in range(ph2_len):
-            s_copy['DeltaPh2',ph2_index] = s_copy['DeltaPh2',ph2_index].run(lambda x,
-                    axis=None: roll(x,ph2_index,axis=axis),'ph2')
+        if ODNP:
+            for ph1_index in range(ph1_len):
+                s_copy['DeltaPh1',ph1_index] = s_copy['DeltaPh1',ph1_index].run(lambda x, 
+                    axis=None: roll(x, ph1_index,axis=axis),'ph1')
+        else:
+            for ph1_index in range(ph1_len):
+                print("PH1 INDEX IS:",ph1_index)
+                s_copy['DeltaPh1',ph1_index] = s_copy['DeltaPh1',ph1_index].run(lambda x, 
+                    axis=None: roll(x, ph1_index,axis=axis),'ph1')
+            for ph2_index in range(ph2_len):
+                s_copy['DeltaPh2',ph2_index] = s_copy['DeltaPh2',ph2_index].run(lambda x,
+                        axis=None: roll(x,ph2_index,axis=axis),'ph2')
         for j in range(1,N):
             correl += s_copy2 * s_copy.C.run(lambda x, axis=None: roll(x,j,axis=axis),
                 indirect_dim).run(conj)
@@ -114,14 +151,21 @@ def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_titl
                 fl.image(correl.C.setaxis('vd','#').set_units('vd','scan #'),human_units=False)
         correl.ft_clear_startpoints('t2')
         correl.ft('t2', shift=True, pad=2**14)
-        correl.ft(['DeltaPh1','DeltaPh2'])
-        correl = correl['DeltaPh1',ph1_selection]['DeltaPh2',ph2_selection] + correl['DeltaPh1',0]['DeltaPh2',0]
+        if ODNP:
+            correl.ft(['DeltaPh1'])
+            correl = correl['DeltaPh1',ph1_selection] + correl['DeltaPh1',0]
+        else:
+            correl.ft(['DeltaPh1','DeltaPh2'])
+            correl = correl['DeltaPh1',ph1_selection]['DeltaPh2',ph2_selection] + correl['DeltaPh1',0]['DeltaPh2',0]
         if my_iter ==0:
             logging.info(strm("holder"))
             if fl is not None:
                 fl.next('correlation function \nfreq domain, after apod')
                 fl.image(correl.C.setaxis('vd','#').set_units('vd','scan #'),human_units=False)
-        f_shift = correl.run(real).argmax('t2')
+        if shift_bounds:
+            f_shift = correl['t2':(-max_shift,max_shift)].run(real).argmax('t2')
+        if not shift_bounds:
+            f_shift = correl.run(real).argmax('t2')
         s_copy = s.C
         s_copy *= exp(-1j*2*pi*f_shift*s_copy.fromaxis('t2'))
         s.ft('t2')
@@ -149,5 +193,11 @@ def correl_align(s, align_phases=False,tol=1e-4,indirect_dim='indirect',fig_titl
     if fl is not None:
         fl.next('after correlation\nph0 restored sig. energy=%g'%sig_energy)
         fl.image(s_copy.C.setaxis('vd','#').set_units('vd','scan #'),human_units=False)
-    return f_shift,sigma    
+    s.ift('t2')
+    s *= np.exp(-1j*2*pi*f_shift*s.fromaxis('t2'))
+    s.ft('t2')
+    if 'nScans' in s.dimlabels:
+        s.chunk('nScans',['nScans','power'],[nScans_len,-1])
+        s.reorder(['ph1','nScans','power','t2'])
+    return s,f_shift,sigma
